% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.usethis-package.R
\docType{package}
\name{r2dii.usethis-package}
\alias{r2dii.usethis}
\alias{r2dii.usethis-package}
\title{r2dii.usethis: Automate Package And Project Setup}
\description{
Automate package and project setup. For more
    information visit <https://2degrees-investing.org/>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/r2dii.usethis}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/r2dii.usethis/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

Authors:
\itemize{
  \item Pranav Pandya \email{pranav.pandya84@yahoo.com} (\href{https://orcid.org/0000-0002-6462-6905}{ORCID})
}

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
