% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data_raw.R
\name{source_data_raw}
\alias{source_data_raw}
\title{Source all .R files under data-raw/}
\usage{
source_data_raw(path = "data-raw")
}
\arguments{
\item{path}{String. Path in the working directory.}
}
\value{
\code{invisible(path)}, as it's called for its side effect.
}
\description{
Usually we work on one dataset only, and don't know if our change impacted
other datasets. This function helps "refresh" all datasets at once. It may
be used interactively while developing the package, or in CI to regularly
check we can reproduce all datasets we export, and that the result is
consistent with our regression tests.
}
\examples{
source_data_raw()
}
