% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_classification.R
\name{use_classification}
\alias{use_classification}
\title{Use a new sector classification in r2dii.data}
\arguments{
\item{prefix}{String. The prefix before _classification, e.g. "psic" in
\code{psic_classification}.}

\item{dataset}{String. Name of the new classification dataset you want to
add, e.g. \code{psic_classification}.}

\item{contributor}{String. Name of a contributor to thank in NEWS.md, e.g.
"\@daisy-pacheco".}

\item{issue}{String. Number of related issue or PR, e.g. #199.}

\item{data}{A data frame.}

\item{string}{A string.}
}
\value{
Most functions are interactive, and called for their side effects.
The return value is usually \code{invisible(dataset)}.
}
\description{
These functions help you add a new sector classification bridge to the
package r2dii.data.
}
\examples{
# See https://github.com/2DegreesInvesting/r2dii.data/issues/199
# Change directory into r2dii.data
\dontrun{
prefix <- "psic"
dataset <- name_dataset(prefix)
contributor <- "@daisy-pacheco"
issue <- "#199"
modify_news(dataset, contributor, issue)
modify_r_classification_bridge(dataset)

x <- readr::read_tsv("~/Downloads/NSCB_PACTA_bridge - PSIC_PACTA_bridge.tsv")
write_raw_data(x, dataset)
modify_data_raw_classification_bridge(dataset)
new_data_dictionary_entry(dataset)
modify_test_data_dictionary(dataset)
}
}
