% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_data_format.R
\name{use_data_format}
\alias{use_data_format}
\title{Document datasets in r2dii.data::data_dictionary}
\usage{
use_data_format(dataset)
}
\arguments{
\item{dataset}{Character string of lentgh-1 giving the name of a dataset
in the column \code{datasets} of \link[r2dii.data:data_dictionary]{r2dii.data::data_dictionary}.}
}
\value{
Text output of class "glue".
}
\description{
Document datasets in r2dii.data::data_dictionary
}
\examples{
use_data_format("region_isos")
}
