% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_logo.R
\name{use_logo}
\alias{use_logo}
\title{Output 2DII's blue logo on a white background.}
\source{
Uploaded to imgur from the file "2DII-Logo-Blue-1000px.png" accessed
from 2DII's dropbox folder on 2020-04-02.
}
\usage{
use_logo(package)
}
\arguments{
\item{package}{String of length-1 giving the name of the package to which you
want to add the logo.}
}
\value{
\code{NULL} as it is called for it's side effect.
}
\description{
Output 2DII's blue logo on a white background.
}
\examples{
use_logo("packagename")
}
