% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pkg_actions.R
\name{use_pkg_actions}
\alias{use_pkg_actions}
\title{Use GitHub actions for packages}
\usage{
use_pkg_actions()
}
\value{
Invisible \code{NULL}.
}
\description{
This function configures GitHub actions to check your package, build its
website, and deploy it to gh-pages. By default, examples don't run. This is
to avoid accidentally showing private data in examples. If you do want to
run examples, edit ".github/workflows/main.yml" and replace
\code{pkgdown::build_site(examples = FALSE)} with
\code{pkgdown::build_site(examples = TRUE)}.
}
\section{Setup}{

\itemize{
\item Setup secrets at Settings > Secrets (available only if you have access to GitHub Actions):
\itemize{
\item You need \code{GH_PAT} for GitHub actions
(\url{https://github.com/maxheld83/ghpages#secrets}).
\item If your package needs to access private repos, also add your
\code{GITHUB_PAT}.
}
\item Use pkgdown with \code{\link[usethis:use_pkgdown]{usethis::use_pkgdown()}} or from 2dii's template:
}\preformatted{usethis::use_template("_pkgdown.yml", package = "r2dii.usethis")
}
\itemize{
\item Create the gh-pages branch from the terminal (instructions from \code{\link[pkgdown:deploy_site_github]{pkgdown::deploy_site_github()}}):
}\preformatted{git checkout --orphan gh-pages
git rm -rf .
git commit --allow-empty -m 'Initial gh-pages commit'
git push origin gh-pages
git checkout master
}
}

\examples{
\dontrun{
use_pkg_actions()
}
}
\seealso{
\code{\link[usethis:use_pkgdown]{usethis::use_pkgdown()}}, \code{\link[pkgdown:deploy_site_github]{pkgdown::deploy_site_github()}}.
}
