% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r2dii_data.R
\name{use_r2dii_data}
\alias{use_r2dii_data}
\title{A checklist to add new data to the package r2dii.data}
\usage{
use_r2dii_data(name = NULL)
}
\arguments{
\item{name}{String of length 1 giving the name of the dataset, e.g.
"ald_demo".}
}
\value{
Printed output that you may copy and paste into a GitHub issue.
}
\description{
A checklist to add new data to the package r2dii.data
}
\examples{
# Maybe copy and paste into a GitHub issue. 
# See `?usethis::browse_github_issues()`
use_r2dii_data()

use_r2dii_data("my_data")
}
