% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_r2dii_labels.R
\name{use_r2dii_labels}
\alias{use_r2dii_labels}
\title{Add default r2dii github issue labels to the active repository}
\usage{
use_r2dii_labels(delete_default = FALSE)
}
\arguments{
\item{delete_default}{If \code{TRUE}, removes GitHub default labels that do not
appear in the \code{labels} vector and that do not have associated issues.}
}
\description{
A slim wrapper around \code{usethis::use_github_labels} to set standard labels
and colours to any \code{r2dii} type github repository. I vaguely followed the
style-guide found \href{https://robinpowered.com/blog/best-practice-system-for-organizing-and-tagging-github-issues}{here}.
}
\examples{
# not run:
dontrun{
  use_r2dii_labels()
}
}
