% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bridge.R
\name{use_bridge}
\alias{use_bridge}
\title{Use a new sector-classification bridge in r2dii.data}
\usage{
use_bridge(dataset, data, contributor = NULL, issue = NULL, overwrite = FALSE)
}
\arguments{
\item{dataset}{String. Name of the new classification dataset you want to
add, with the format \verb{[prefix]_classification}, e.g. \code{psic_classification}.}

\item{data}{A data frame.}

\item{contributor}{String. Name of a contributor to thank in NEWS.md, e.g.
"\@daisy-pacheco".}

\item{issue}{String. Number of related issue or PR, e.g. #199.}

\item{overwrite}{Logical. Allow overwriting data?}
}
\value{
Most functions are interactive, and called for their side effects.
The return value is usually \code{invisible(dataset)}.
}
\description{
Use a new sector-classification bridge in r2dii.data
}
\examples{
\dontrun{
library(r2dii.usethis)

fs::dir_copy("../r2dii.data", tempdir())
tmp <- fs::path(tempdir(), "r2dii.data")
old <- getwd()

setwd(tmp)

# This is usually a contributed spreadsheet
dataset <- "fake"
data <- data.frame(
  original_code = letters[1:3],
  code_level = 5,
  code = as.character(1:3),
  sector = "not in scope",
  borderline = FALSE
)
contributor <- "@somebody"
issue <- "#123"
use_bridge(dataset, data, contributor, issue)

system("git status -s")

# Teardown
unlink(tmp, recursive = TRUE)
setwd(old)
}
}
