% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bridge.R
\name{use_bridge}
\alias{use_bridge}
\alias{bridge_update_news}
\alias{name_dataset}
\alias{bridge_update_r}
\alias{bridge_write_raw_data}
\alias{bridge_update_data_raw}
\alias{bridge_add_dictionary}
\alias{bridge_update_snapshots}
\title{Use a new sector classification in r2dii.data}
\usage{
use_bridge(dataset, data, contributor = NULL, issue = NULL, overwrite = FALSE)

bridge_update_news(dataset, contributor = NULL, issue = NULL)

name_dataset(prefix)

bridge_update_r(dataset)

bridge_write_raw_data(dataset, data, overwrite = FALSE)

bridge_update_data_raw(dataset)

bridge_add_dictionary(dataset, overwrite = FALSE)

bridge_update_snapshots()
}
\arguments{
\item{dataset}{String. Name of the new classification dataset you want to
add, with the format \verb{[prefix]_classification}, e.g. \code{psic_classification}.}

\item{data}{A data frame.}

\item{contributor}{String. Name of a contributor to thank in NEWS.md, e.g.
"\@daisy-pacheco".}

\item{issue}{String. Number of related issue or PR, e.g. #199.}

\item{overwrite}{Logical. Allow overwriting data?}
}
\value{
Most functions are interactive, and called for their side effects.
The return value is usually \code{invisible(dataset)}.
}
\description{
These functions help you add a new sector classification bridge.
}
\examples{
\dontrun{
# This is an internal function aimed at developers
load_all()

# The source code does not ship with the package to avoid dependencies
source_use_bridge()

# This is usually a contributed spreadsheet
data <- r2dii.data::psic_classification

# The dataset name must have the format [prefix]_classificaton
dataset <- name_dataset("fake")
dataset

contributor <- "@somebody"
issue <- "#123"

use_bridge(dataset, data, contributor, issue, overwrite = TRUE)
}
}
