% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_column_definitions.R
\name{use_column_definitions}
\alias{use_column_definitions}
\title{Define columns of a dataset}
\usage{
use_column_definitions(dictionary, dataset)
}
\arguments{
\item{dictionary}{A dataframe structured as in the example below.}

\item{dataset}{A character string giving the name of a dataset.}
}
\value{
A character vector of subclass "glue".
}
\description{
This function helps you write the \verb{@format} section of a dataset's help file.
Write the help file with that packages roxygen2 and call this function with:\preformatted{`r r2dii.usethis::use_column_definitions()`
}
}
\examples{
dictionary <- data.frame(
  stringsAsFactors = FALSE,
  dataset = paste0("dataset_", c(1, 1, 2, 2)),
  column = letters[1:4],
  typeof = c("character", "integer", "double", "factor"),
  definition = paste0("Definition of ", letters[1:4])
)

dictionary 


use_column_definitions(dictionary, dataset = "dataset_1")

use_column_definitions(dictionary, dataset = "dataset_2")
}
